Exception refactor
